requirements.txt
streamlit
langchain
langchain-groq
pypdf
python-docx
pandas
sentence-transformers
faiss-cpu
wikipedia
python-dotenv
langchain-community